requirements.txt

boto3
kafka-python
pyspark
PyYAML
psycopg2-binary
